requirements.txt
gradio==5.20.1
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2